--- alembic/old/158d419ae542_definition.py
@@ -1,54 +1,0 @@
-"""definition
-
-Revision ID: 158d419ae542
-Revises: e601685c1897
-Create Date: 2021-01-19 23:31:01.867461
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '158d419ae542'
-down_revision = 'e601685c1897'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('Definition',
-    sa.Column('Id', sa.Integer(), nullable=False),
-    sa.Column('Name', sa.String(length=100), nullable=False),
-    sa.Column('Version', sa.Integer(), nullable=False),
-    sa.Column('Content', sa.Text(), nullable=True),
-    sa.Column('CreatedByUserId', sa.Integer(), nullable=False),
-    sa.Column('CreationDate', sa.DateTime(), nullable=False),
-    sa.Column('LastUpdatedUserId', sa.Integer(), nullable=True),
-    sa.Column('LastUpdatedDate', sa.DateTime(), nullable=True),
-    sa.Column('IsDeleted', sa.Integer(), nullable=False),
-    sa.Column('Comments', sa.String(length=1000), nullable=True),
-    sa.Column('RowVersion', sa.TIMESTAMP(), nullable=True),
-    sa.PrimaryKeyConstraint('Id'),
-    schema='Operation'
-    )
-    op.add_column('DataIntegration', sa.Column('DefinitionId', sa.Integer(), nullable=True), schema='Integration')
-    op.create_foreign_key(None, 'DataIntegration', 'Definition', ['DefinitionId'], ['Id'], source_schema='Integration', referent_schema='Operation')
-    op.add_column('DataOperation', sa.Column('DefinitionId', sa.Integer(), nullable=True), schema='Operation')
-    op.create_foreign_key(None, 'DataOperation', 'Definition', ['DefinitionId'], ['Id'], source_schema='Operation', referent_schema='Operation')
-    op.add_column('DataOperationJobExecution', sa.Column('DefinitionId', sa.Integer(), nullable=True), schema='Operation')
-    op.create_foreign_key(None, 'DataOperationJobExecution', 'Definition', ['DefinitionId'], ['Id'], source_schema='Operation', referent_schema='Operation')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'DataOperationJobExecution', schema='Operation', type_='foreignkey')
-    op.drop_column('DataOperationJobExecution', 'DefinitionId', schema='Operation')
-    op.drop_constraint(None, 'DataOperation', schema='Operation', type_='foreignkey')
-    op.drop_column('DataOperation', 'DefinitionId', schema='Operation')
-    op.drop_constraint(None, 'DataIntegration', schema='Integration', type_='foreignkey')
-    op.drop_column('DataIntegration', 'DefinitionId', schema='Integration')
-    op.drop_table('Definition', schema='Operation')
-    # ### end Alembic commands ###
