--- alembic/old/c207709fadb3_config_parameter.py
@@ -1,43 +1,0 @@
-"""config_parameter
-
-Revision ID: c207709fadb3
-Revises: f49062fc3aee
-Create Date: 2021-01-10 15:59:18.257057
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'c207709fadb3'
-down_revision = 'f49062fc3aee'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('ConfigParameter',
-    sa.Column('Id', sa.Integer(), nullable=False),
-    sa.Column('Name', sa.String(length=255), nullable=False),
-    sa.Column('Type', sa.String(length=255), nullable=True),
-    sa.Column('Value', sa.String(length=255), nullable=False),
-    sa.Column('Description', sa.String(length=1000), nullable=False),
-    sa.Column('CreatedByUserId', sa.Integer(), nullable=False),
-    sa.Column('CreationDate', sa.DateTime(), nullable=False),
-    sa.Column('LastUpdatedUserId', sa.Integer(), nullable=True),
-    sa.Column('LastUpdatedDate', sa.DateTime(), nullable=True),
-    sa.Column('IsDeleted', sa.Integer(), nullable=False),
-    sa.Column('Comments', sa.String(length=1000), nullable=True),
-    sa.Column('RowVersion', sa.TIMESTAMP(), nullable=True),
-    sa.PrimaryKeyConstraint('Id'),
-    schema='Common'
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('ConfigParameter', schema='Common')
-    # ### end Alembic commands ###
