--- alembic/old/e601685c1897_data_operation_job_execution_integration.py
@@ -1,70 +1,0 @@
-"""data_operation_job_execution_integration
-
-Revision ID: e601685c1897
-Revises: 8fc05e3e926b
-Create Date: 2021-01-19 22:41:02.537477
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'e601685c1897'
-down_revision = '8fc05e3e926b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('DataOperationJobExecutionIntegration',
-    sa.Column('Id', sa.Integer(), nullable=False),
-    sa.Column('DataOperationJobExecutionId', sa.Integer(), nullable=True),
-    sa.Column('DataOperationIntegrationId', sa.Integer(), nullable=True),
-    sa.Column('StatusId', sa.Integer(), nullable=True),
-    sa.Column('StartDate', sa.DateTime(), nullable=False),
-    sa.Column('EndDate', sa.DateTime(), nullable=True),
-    sa.Column('Limit', sa.Integer(), nullable=True),
-    sa.Column('ProcessCount', sa.Integer(), nullable=True),
-    sa.Column('SourceDataCount', sa.Integer(), nullable=True),
-    sa.Column('Log', sa.String(length=1000), nullable=True),
-    sa.Column('CreatedByUserId', sa.Integer(), nullable=False),
-    sa.Column('CreationDate', sa.DateTime(), nullable=False),
-    sa.Column('LastUpdatedUserId', sa.Integer(), nullable=True),
-    sa.Column('LastUpdatedDate', sa.DateTime(), nullable=True),
-    sa.Column('IsDeleted', sa.Integer(), nullable=False),
-    sa.Column('Comments', sa.String(length=1000), nullable=True),
-    sa.Column('RowVersion', sa.TIMESTAMP(), nullable=True),
-    sa.ForeignKeyConstraint(['DataOperationIntegrationId'], ['Operation.DataOperationIntegration.Id'], ),
-    sa.ForeignKeyConstraint(['DataOperationJobExecutionId'], ['Operation.DataOperationJobExecution.Id'], ),
-    sa.ForeignKeyConstraint(['StatusId'], ['Common.Status.Id'], ),
-    sa.PrimaryKeyConstraint('Id'),
-    schema='Operation'
-    )
-    op.create_table('DataOperationJobExecutionIntegrationEvent',
-    sa.Column('Id', sa.Integer(), nullable=False),
-    sa.Column('DataOperationJobExecutionIntegrationId', sa.Integer(), nullable=True),
-    sa.Column('EventId', sa.Integer(), nullable=True),
-    sa.Column('EventDate', sa.DateTime(), nullable=False),
-    sa.Column('AffectedRowCount', sa.Integer(), nullable=True),
-    sa.Column('CreatedByUserId', sa.Integer(), nullable=False),
-    sa.Column('CreationDate', sa.DateTime(), nullable=False),
-    sa.Column('LastUpdatedUserId', sa.Integer(), nullable=True),
-    sa.Column('LastUpdatedDate', sa.DateTime(), nullable=True),
-    sa.Column('IsDeleted', sa.Integer(), nullable=False),
-    sa.Column('Comments', sa.String(length=1000), nullable=True),
-    sa.Column('RowVersion', sa.TIMESTAMP(), nullable=True),
-    sa.ForeignKeyConstraint(['DataOperationJobExecutionIntegrationId'], ['Operation.DataOperationJobExecutionIntegration.Id'], ),
-    sa.ForeignKeyConstraint(['EventId'], ['Common.OperationEvent.Id'], ),
-    sa.PrimaryKeyConstraint('Id'),
-    schema='Operation'
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('DataOperationJobExecutionIntegrationEvent', schema='Operation')
-    op.drop_table('DataOperationJobExecutionIntegration', schema='Operation')
-    # ### end Alembic commands ###
