--- alembic/old/f49062fc3aee_integration_query.py
@@ -1,28 +1,0 @@
-"""integration_query
-
-Revision ID: f49062fc3aee
-Revises: b4b374eb4b81
-Create Date: 2020-12-26 20:35:49.684153
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'f49062fc3aee'
-down_revision = 'b4b374eb4b81'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('DataIntegrationConnection', sa.Column('Query', sa.Text(), nullable=True), schema='Integration')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('DataIntegrationConnection', 'Query', schema='Integration')
-    # ### end Alembic commands ###
