--- alembic/update.db
@@ -1,48 +1,0 @@
-ALTER TABLE "Integration"."PythonDataIntegration" RENAME TO "DataIntegration";
-ALTER TABLE "Integration"."PythonDataIntegrationColumn" RENAME TO "DataIntegrationColumn";
-ALTER TABLE "Integration"."PythonDataIntegrationExecutionJob" RENAME TO "DataIntegrationExecutionJob";
-ALTER TABLE "Integration"."PythonDataIntegrationConnection" RENAME TO "DataIntegrationConnection";
-ALTER TABLE "Integration"."PythonDataIntegrationLog" RENAME TO "Log";
-DROP TABLE "Integration"."PythonDataIntegrationJob";
-ALTER TABLE "Integration"."DataIntegrationColumn"
-    RENAME "PythonDataIntegrationId" TO "DataIntegrationId";
-ALTER TABLE "Integration"."DataIntegration" 
-RENAME CONSTRAINT "PythonDataIntegration_pkey" TO "DataIntegration_pkey";
-ALTER INDEX "Integration"."ix_Integration_PythonDataIntegration_Code"
-    RENAME TO "ix_Integration_DataIntegration_Code";
-ALTER TABLE "Integration"."DataIntegrationColumn"
-RENAME CONSTRAINT "PythonDataIntegrationColumn_PythonDataIntegrationId_fkey" TO "DataIntegrationColumn_DataIntegrationId_fkey";
-ALTER TABLE "Integration"."DataIntegrationColumn"
-    RENAME CONSTRAINT "PythonDataIntegrationColumn_pkey" TO "DataIntegrationColumn_pkey";
-ALTER TABLE "Integration"."DataIntegrationConnection"
-RENAME "PythonDataIntegrationId" TO "DataIntegrationId";
-ALTER TABLE "Integration"."DataIntegrationConnection"
-    RENAME CONSTRAINT "PythonDataIntegrationConnection_ConnectionId_fkey" TO "DataIntegrationConnection_ConnectionId_fkey";
-ALTER TABLE "Integration"."DataIntegrationConnection"
-    RENAME CONSTRAINT "PythonDataIntegrationConnection_PythonDataIntegrationId_fkey" TO "DataIntegrationConnection_DataIntegrationId_fkey";
-	ALTER TABLE "Integration"."DataIntegrationConnection"
-    RENAME CONSTRAINT "PythonDataIntegrationConnection_pkey" TO "DataIntegrationConnection_pkey";
-	ALTER TABLE "Integration"."DataIntegrationExecutionJob"
-    RENAME "PythonDataIntegrationId" TO "DataIntegrationId";
-	ALTER TABLE "Integration"."DataIntegrationExecutionJob"
-    RENAME CONSTRAINT "PythonDataIntegrationExecutionJob_PythonDataIntegrationId_fkey" TO "DataIntegrationExecutionJob_DataIntegrationId_fkey";
-	ALTER TABLE "Integration"."DataIntegrationExecutionJob"
-    RENAME CONSTRAINT "PythonDataIntegrationExecutionJob_pkey" TO "DataIntegrationExecutionJob_pkey";
-	ALTER TABLE "Integration"."Log"
-    RENAME CONSTRAINT "PythonDataIntegrationLog_pkey" TO "DataIntegrationLog_pkey";
-	ALTER INDEX "Integration"."ix_Integration_PythonDataIntegrationLog_JobId"
-    RENAME TO "ix_Integration_DataIntegrationLog_JobId";
-	ALTER INDEX "Integration"."ix_Integration_PythonDataIntegrationLog_TypeId"
-    RENAME TO "ix_Integration_DataIntegrationLog_TypeId";
-	ALTER TABLE "Operation"."DataOperationIntegration"
-    RENAME CONSTRAINT "DataOperationIntegration_PythonDataIntegrationId_fkey" TO "DataOperationIntegration_DataIntegrationId_fkey";
-ALTER TABLE "Operation"."DataOperationIntegration"
-RENAME "PythonDataIntegrationId" TO "DataIntegrationId";
-	ALTER TABLE "Integration"."Log"
-  SET SCHEMA "Common";
-  ALTER INDEX "Common"."ix_Integration_DataIntegrationLog_JobId"
-    RENAME TO "ix_Common_Log_JobId";
-ALTER INDEX "Common"."ix_Integration_DataIntegrationLog_TypeId"
-RENAME TO "ix_Common_Log_TypeId";
-ALTER TABLE "Connection"."ConnectionDatabase"
-    ADD COLUMN "ServiceName" character varying(100);
