+++ application.test.yml
@@ -2,19 +2,11 @@
 DATABASE:
   #POSTGRE
   TYPE: POSTGRESQL
-  HOST: 10.220.13.77
+  HOST: localhost
   PORT: 5432
-  DATABASE: pdi
-  USERNAME: pdi_user
-  PASSWORD: PWD_011
-  SCHEMA: Common
-#  TYPE: POSTGRESQL
-#  HOST: localhost
-#  PORT: 5432
-#  DATABASE: test
-#  USERNAME: postgres
-#  PASSWORD: 123456
-#  SCHEMA: Common
+  DATABASE: test
+  USERNAME: postgres
+  PASSWORD: 123456
 
 API:
   ENVIRONMENT: DEVELOPMENT
