+++ application.yml
@@ -3,7 +3,7 @@ DATABASE:
   TYPE: POSTGRESQL
   HOST: localhost
   PORT: 5432
-  DATABASE: test
+  DATABASE: test_pdi
   USERNAME: postgres
   PASSWORD: 123456
