+++ models/enums/events.py
@@ -2,3 +2,9 @@
 EVENT_EXECUTION_INITIALIZED=1
 EVENT_EXECUTION_STARTED=2
 EVENT_EXECUTION_FINISHED=3
+EVENT_EXECUTION_INTEGRATION_STARTED=4
+EVENT_EXECUTION_INTEGRATION_EXECUTE_PRE_PROCEDURE=5
+EVENT_EXECUTION_INTEGRATION_EXECUTE_POST_PROCEDURE=6
+EVENT_EXECUTION_INTEGRATION_EXECUTE_TRUNCATE=7
+EVENT_EXECUTION_INTEGRATION_EXECUTE_QUERY=8
+EVENT_EXECUTION_INTEGRATION_EXECUTE_OPERATION=9
